Fix nested NavItems
Currently, a NavItem having more than one child will throw an error. This is due to the `children` property of `NavItem` being passed as the first argument to `React.cloneElement`, which expects a single element. The existing tests are only verifying the case where the `NavItem` has only one child.

This commit maps over `children` and calls `React.cloneElement` on each of them.