Techdocs-cli Build Fix
We're using cloud build in gcp to build/package our backstage deployment
and when you upload your artifacts it will remove all of the .git folder
to mimimize the upload causing the techdocs-cli to fail as it relies on
using `git rev-parse` to discover the full path. I changed this to be
statically linked so the reliance on git is no longer there and so that
as long as the structure of your repo doesn't change it should build no
matter where the root folder is placed.