Fix view for Table of Contents on smaller screens
For backstage.io/docs and backstage.io/blog pages, we have a sidebar with Table of Contents. On smaller screens, this sidebar becomes a nav bar on top with a toggle hamburger button. However, we also have a nav bar on the website. This nav bar hinders the view of the TOC bar.

This commit fixes the view for Table of Contents on Blog and Docs page on smaller screens.