Use app-config.development.yaml to provide configuration instead of local
The app-config.yaml file that comes inside the docker image for the
backend contains configuration for the key `scaffolder.azure`. This
automatically enables the use of azure and it is not possible to set its
values to dummy ones because it will fail when it tries to validate
them.

To avoid the backend throwing an exception for this, the charts are
using the "development" mode (via setting NODE_ENV=development).

This means that the app-config.development.yaml file that ships in the
container woud have higher precedence than the app-config.local.yaml
file that the charts tried to create before. For that reason, for now,
they will override the development file.

A better solution would be to ship the docker images with a minimal
app-config.yaml or even without any config. And then have these charts
provide said file.