Fix missing authentication in GitHubUrlReader
An oversight that probably stems from that most repositories for tech-docs are public anyway. 
Ours are not, and are inside of a GitHub Enterprise, which is how i found this tiny bug. 

I wish i knew how to mock the cross-fetch in order to inspect the Authorization header of the request to test this fix. 
I can only confirm that it works on our backstage that targets GHE.